[move] move 'cache' items into voip items
